virtio-balloon: Add free page hinting
Free page hinting is a mechanism which allows the guest driver to report
ranges of pages to the host device. A "hinting" run is triggered by the
device by issuing a new command id in the config space, after the update
to the id the device will hint ranges to the host which are unused. Once
the driver has exhausted all free ranges it notifies the device the run
has completed. The device can then issue another command allowing the
guest to reclaim these pages.

Adding support for hinting the firecracker balloon device, we offer
three points to manage the device; first to start a run, second to
monitor the status and a final to issue the command to allow the guest
to reclaim pages.

To note, there is a potential condition in the linux driver which would
allow a range to be reclaimed in an oom scenario before we remove the
range.

Signed-off-by: Jack Thomson <jackabt@amazon.com>