test: Add integration tests for hinting/reporting
Add integration tests for free page hinting and reporting, both
functional and performance tests.

Update fast_page_helper so it can run in a oneshot mode, not requiring
the signal to track the performance.

New functional tests to ensure that hinting and reporting are reducing
the RSS as expected in the guest. Updated reduce RSS test to touch
memory to reduce the chance of flakiness.

New performance tests for the balloon device. First being a test to
track the CPU overhead of hinting and reporting. Second being a test to
measure the faulting latency while reporting is running in the guest.

Signed-off-by: Jack Thomson <jackabt@amazon.com>